pyproject.toml
[project]
name = "octotail"
dynamic = ["version"]
description = "Live tail GitHub Actions runs on git push"
authors = [{ name = "Rareș Cosma", email = "rares@getbetter.ro" }]
readme = "README.md"
requires-python = ">=3.12"
keywords = ["github-actions", "tail", "post-receive", "codecrafters", "git"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Version Control :: Git",
  "Topic :: System :: Logging",
  "License :: OSI Approved :: The Unlicense (Unlicense)",
]
license = { text = "Unlicense" }
dependencies = [
  "fake-useragent>=1.5",
  "mitmproxy>=11.0",
  "pygithub>=2.4",
  "pykka>=4.0",
  "pyppeteer>=2.0",
  "pyppeteer-stealth>=2.7",
  "pyxdg>=0.28",
  "returns>=0.21",
  "rich>=13.9",
  "shellingham>=1.5",
  "termcolor>=2.5",
  "typer>=0.12",
  "websockets>=10",
]

[project.scripts]
octotail = "octotail.main:_main"
octotailx = "octotail.x:_main"

[project.urls]
Documentation = "https://github.com/rarescosma/octotail#readme"
Repository = "https://github.com/rarescosma/octotail"
Issues = "https://github.com/rarescosma/octotail/issues"
Homepage = "https://getbetter.ro/projects/#octotail"

[tool.uv]
dev-dependencies = [
  "black>=24.10",
  "bumpversion>=0.6",
  "coverage[toml]>=7.6.4",
  "isort>=5.13",
  "mypy>=1.12",
  "ptyprocess>=0.7",
  "pycln>=2.4",
  "pylint>=3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "pytest>=8.3.3",
  "returns[compatible-mypy]>=0.21.0",
  "ruff>=0.6.9",
  "vulture>=2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["/octotail", "/examples"]

[tool.hatch.version]
path = "octotail/__init__.py"

[tool.ruff]
line-length = 100
lint.extend-select = [
  "B",
  "C",
  "E",
  "F",
  "I",
  "PERF",
  "PLR",
  "PT",
  "PTH",
  "RUF",
  "SIM",
  "TID",
  "TRY",
  "UP",
  "W",
]
lint.ignore = ["B010", "C901", "COM", "D", "E501", "TRY003", "PLR0912"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true
source = ["octotail"]
omit = ["main.py", "x.py"]

[tool.coverage.report]
show_missing = true
